+++ src/controllers/controllers/planner.py
@@ -5,6 +5,7 @@ import heapq
 from .path_planners.DStarLite import DStarLite
 from .path_planners.PriorityQueue import PriorityQueue
 from .path_planners.util import heuristic
+from.path_planners.ImageGraph import ImageGraph
 
 
 def main():
@@ -139,23 +140,7 @@ def main():
     # Close the CSV file
     file.close()
 
-class ImageGraph:
-    def __init__(self, image):
-        self.image = image
-    
-    def neighbors(self, id):
-        x, y = id
-        neighbors = []
-        for dx, dy in [(-1, 0), (1, 0), (0, -1), (0, 1)]:
-            nx, ny = x + dx, y + dy
-            if 0 <= nx < self.image.shape[0] and 0 <= ny < self.image.shape[1] and self.image[nx, ny] == 255:
-                neighbors.append((nx, ny))
-        return neighbors
-    
-    def cost(self, from_id, to_id):
-        # Here, you could use a simple distance metric (like Euclidean or Manhattan distance)
-        # Calculate Euclidean distance between from_id and to_id
-        return np.sqrt((from_id[0] - to_id[0]) ** 2 + (from_id[1] - to_id[1]) ** 2)
+
 
 def a_star(image, start, goal):
     graph = ImageGraph(image)
